pyproject.toml
[tool.poetry]
name = "barbarzyncy"
version = "0.1.0"
authors = ["Lukasz Zmudzinski <lukasz@zmudzinski.me>"]
description = "Website for Barbarzyncy guild on Burning Legion"

[tool.poetry.dependencies]
python = "<4.0,>=3.9"
Django = "~3.2"
django-cors-headers = "^3.7.0"
django-environ = "^0.4.5"
django-storages = { version = "^1.11.1", extras = ["boto3"] }
pytz = "^2021.1"
pyuwsgi = "^2.0.20"
psycopg2-binary = {version = "^2.9.1", optional = true}
requests = "^2.27.1"
types-requests = "^2.27.16"
django-tinymce = "^3.4.0"
whitenoise = "^6.0.0"
python-dotenv = "^0.20.0"
Pillow = "^9.1.0"
mysqlclient = "^2.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django-debug-toolbar = "^3.2"
django-extensions = "^3.1.2"
flake8 = "^3.9.0"
flake8-blind-except = "^0.2.0"
flake8-debugger = "^4.0.0"
flake8-isort = "^4.0.0"
flake8-print = "^4.0.0"
flake8-quotes = "^3.2.0"
factory-boy = "^3.2.0"
ipdb = "^0.13.7"
ipython = "^7.22.0"
isort = "^5.8.0"
mypy = "^0.910"
pep8-naming = "^0.11.1"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
pytest-django = "^4.2.0"
pytest-freezegun = "^0.4.2"

[tool.poetry.extras]
prod = ["psycopg2"]
dev = ["psycopg2-binary"]

[tool.isort]
line_length = 88
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
balanced_wrapping = true
skip = "manage.py, migrations"
known_django = "django"
sections = "FUTURE, STDLIB, DJANGO, THIRDPARTY, FIRSTPARTY, LOCALFOLDER"

[tool.coverage.run]
source = ["/project/app/src/"]
omit = [
  "*apps.py",
  "*manage.py",
  "*commands/bootstrap.py",
  "*migrations/*",
  "*settings*",
  "*urls.py",
  "*wsgi.py"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py *_tests.py"
addopts = ["--nomigrations", "--cov=src", "--cov-report=term-missing:skip-covered",
  "--cov-report=xml:test-results/coverage.xml", "--junitxml=test-results/junit-report.xml"]
testpaths = "/project/app/tests/"
python_classes = "*Test"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
